+++ src/SFA.DAS.ApprenticeAccounts.Jobs/Api/Apprentice.cs
@@ -1,12 +1,15 @@
 ﻿using System;
+using SFA.DAS.ApprenticePortal.Authentication;
 
 namespace SFA.DAS.ApprenticeAccounts.Jobs.Api
 {
-    public class Apprentice
+    public class Apprentice : IApprenticeAccount
     {
-        public Guid Id { get; set; }
+        public Guid ApprenticeId { get; set; }
         public string FirstName { get; set; } = null!;
         public string LastName { get; set; } = null!;
         public string Email { get; set; } = null!;
+        public DateTime DateOfBirth { get; set; }
+        public bool TermsOfUseAccepted { get; set; }
     }
 }
