+++ src/SFA.DAS.ApprenticeAccounts.Jobs/Api/ApprenticeshipConfirmationRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.ApprenticeAccounts.Jobs.Infrastructure.OuterApi
+﻿namespace SFA.DAS.ApprenticeAccounts.Jobs.Api
 {
     public class ApprenticeshipConfirmationRequest
     {
