+++ src/SFA.DAS.ApprenticeAccounts.Jobs/Api/ApprenticeshipDetailsConfirmationRequest.cs
@@ -1,4 +1,4 @@
-﻿    namespace SFA.DAS.ApprenticeCommitments.Web.Services.OuterApi
+﻿namespace SFA.DAS.ApprenticeAccounts.Jobs.Api
 {
     public class ApprenticeshipDetailsConfirmationRequest
     {
