+++ src/SFA.DAS.ApprenticeAccounts.Jobs/Api/ApprenticeshipsWrapper.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace SFA.DAS.ApprenticeAccounts.Jobs.Infrastructure.OuterApi
+namespace SFA.DAS.ApprenticeAccounts.Jobs.Api
 {
     public class ApprenticeshipsWrapper
     {
