+++ src/SFA.DAS.ApprenticeAccounts.Jobs/Api/RolesAndResponsibilitiesConfirmationRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.ApprenticeAccounts.Jobs.Infrastructure.OuterApi
+﻿namespace SFA.DAS.ApprenticeAccounts.Jobs.Api
 {
     public class RolesAndResponsibilitiesConfirmationRequest
     {
