+++ src/SFA.DAS.ApprenticeAccounts.Jobs/Api/VerifyRegistrationResponse.cs
@@ -2,7 +2,7 @@
 
 #nullable disable
 
-namespace SFA.DAS.ApprenticeAccounts.Jobs.Infrastructure.OuterApi
+namespace SFA.DAS.ApprenticeAccounts.Jobs.Api
 {
     public class VerifyRegistrationResponse
     {
