+++ src/SFA.DAS.ApprenticeAccounts.Jobs/Infrastructure/AccountsApiClientConfiguration.cs
@@ -2,7 +2,6 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using RestEase.HttpClientFactory;
 using SFA.DAS.ApprenticeAccounts.Jobs.Api;
-using SFA.DAS.ApprenticeAccounts.Jobs.Infrastructure.OuterApi;
 using SFA.DAS.Http.Configuration;
 using SFA.DAS.Http.TokenGenerators;
 using System;
