--- src/SFA.DAS.ApprenticeAccounts.Jobs/Infrastructure/OuterApi/Apprentice.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using SFA.DAS.ApprenticePortal.Authentication;
-
-namespace SFA.DAS.ApprenticeAccounts.Jobs.Infrastructure.OuterApi
-{
-    public sealed class Apprentice : IApprenticeAccount
-    {
-        public Guid ApprenticeId { get; set; }
-        public string FirstName { get; set; } = null!;
-        public string LastName { get; set; } = null!;
-        public string Email { get; set; } = null!;
-        public DateTime DateOfBirth { get; set; }
-        public bool TermsOfUseAccepted { get; set; }
-    }
-}
